Remove `ActionClickDrill` leftovers
